==> Scripts/2_Emission_Data_Subset.Rmd <==
---
title: "2_Emission_Data_Subset"
output: html_document
date: "2025-09-26"
---

```{r}

merged_metadata <- readRDS("C:/Users/mlolita/OneDrive - CGIAR/Alliance - ClimateActionNetZero - 1_Projects/G227-A1712_SAAF/2_Technical & Data/Lolita Emissions Paper/Emission_Analysis/data/merged_metadata_harmonized.rds")

ingredients <- readRDS("C:/Users/mlolita/OneDrive - CGIAR/Alliance - ClimateActionNetZero - 1_Projects/G227-A1712_SAAF/2_Technical & Data/Lolita Emissions Paper/Emission_Analysis/data/ingredients_harmonized.rds")




```

## 1.4 Take out grazing experiments 

```{r}

species_by_bcode <- merged_metadata$Prod.Out %>%
  select(B.Code, P.Product) %>%
  distinct()

# 2. Papers per species before filtering
# papers_before <- merged_metadata$Animals.Diet %>%
#   select(B.Code) %>%
#   distinct() %>%
#   left_join(species_by_bcode, by = "B.Code") %>%
#   count(P.Product, name = "papers_before")

# Step 1: Identify A.Level.Name (diet) that include Grazing
grazing_diets <- merged_metadata$Animals.Diet %>%
  filter(D.Type == "Grazing") %>%
  distinct(B.Code, A.Level.Name)

length(unique(grazing_diets$B.Code))

# Step 2: Count how many non-grazing diets exist per B.Code
nongrazing_counts <- merged_metadata$Animals.Diet %>%
  filter(!(D.Type == "Grazing")) %>%
  distinct(B.Code, A.Level.Name) %>%
  group_by(B.Code) %>%
  summarise(n_nongrazing = n(), .groups = "drop")

length(unique(nongrazing_counts$B.Code))


# Step 3: Keep only diets that are not marked as Grazing AND are in B.Codes with at least one non-grazing diet
merged_metadata$Animals.Diet <- merged_metadata$Animals.Diet %>%
  anti_join(grazing_diets, by = c("B.Code", "A.Level.Name")) %>%  # remove all diets that were grazing
  semi_join(nongrazing_counts, by = "B.Code")  # keep only papers with at least one non-grazing diet


species_by_bcode <- merged_metadata$Prod.Out %>%
  select(B.Code, P.Product) %>%
  distinct()

# 2. Papers per species before filtering
# papers_after <- merged_metadata$Animals.Diet %>%
#   select(B.Code) %>%
#   distinct() %>%
#   left_join(species_by_bcode, by = "B.Code") %>%
#   count(P.Product, name = "papers_after")

rm(grazing_diets,nongrazing_counts)
```

Subset for Dry data 

When feed intake data is missing and we rely on ingredient-level amounts, we need to ensure that the quantities used refer to the dry form of the ingredient. To do this, we subset the data to include only ingredients identified as dried—based on either the declared DC.Is.Dry value or our earlier flagging using ingredient names and moisture content predictions. This ensures consistency and accuracy in downstream calculations. 
```{r,message=FALSE,warning=FALSE}
#take out unspecified ingredients

species_by_bcode <- merged_metadata$Prod.Out %>%
  select(B.Code, P.Product) %>%
  distinct()

# 2. Papers per species before filtering
# papers_before <- ingredients %>%
#   select(B.Code) %>%
#   distinct() %>%
#   left_join(species_by_bcode, by = "B.Code") %>%
#   count(P.Product, name = "papers_before")
# length(unique(ingredients$B.Code))


# Step 1: Identify A.Level.Name groups with at least one ingredient NOT dry
non_dry_diets <- ingredients %>%
  filter(is.na(DC.Is.Dry) | str_to_lower(DC.Is.Dry) %in% c("no", "n", "unspecified")) %>%
  distinct(B.Code, A.Level.Name)

# Step 2: Filter out these diets entirely from ingredients
ingredients <- ingredients %>%
  anti_join(non_dry_diets, by = c("B.Code", "A.Level.Name"))

length(unique(ingredients$A.Level.Name))

rm(non_dry_diets)

length(unique(ingredients$B.Code))
species_by_bcode <- merged_metadata$Prod.Out %>%
  select(B.Code, P.Product) %>%
  distinct()

#2. Papers per species before filtering
papers_after <- ingredients %>%
  select(B.Code) %>%
  distinct() %>%
  left_join(species_by_bcode, by = "B.Code") %>%
  count(P.Product, name = "papers_after")

```

## 3) Subset data




```{r}


write.csv(ingredients, "C:/Users/mlolita/OneDrive - CGIAR/Alliance - ClimateActionNetZero - 1_Projects/G227-A1712_SAAF/2_Technical & Data/Lolita Emissions Paper/Emission_Analysis/data/ingredients_harmonized_subset.csv")

saveRDS(merged_metadata, "C:/Users/mlolita/OneDrive - CGIAR/Alliance - ClimateActionNetZero - 1_Projects/G227-A1712_SAAF/2_Technical & Data/Lolita Emissions Paper/Emission_Analysis/data/merged_metadata_harmonized_subset.rds")

```

## number of papers 
```{r}

GE_combined <- read.csv("C:/Users/mlolita/OneDrive - CGIAR/Alliance - ClimateActionNetZero - 1_Projects/G227-A1712_SAAF/2_Technical & Data/Lolita Emissions Paper/Emission_Analysis/data/GE_Combined.csv")

GE_combined<-GE_combined%>%
  filter(!is.na(CH4_kg_year))

GE_combined<- GE_combined%>%
  filter(B.Code %in% ingredients$B.Code)

n_CH4_codes_by_product <- GE_combined %>%
  filter(!is.na(CH4_kg_year)) %>%
  distinct(B.Code, P.Product) %>%
  count(P.Product, name = "n_CH4_codes")

n_CH4_codes_by_product

```
